Replace dependency on apacheds-all with narrower-scoped equivalents
The apacheds-all artifact is problematic as it embeds old versions of
a number of other libraries. For example, it contains the SLF4J API,
Antlr, and Bouncycastle. These embedded libraries have not been
repackaged and can therefore cause version clashes. Notably, the
embedded SLF4J API is old (1.5.x) and can cause failures when recent
versions of Logback are also on the classpath. A hacky workaround is to
carefully order the classpath and ensure that a newer version of the
SLF4J API comes before the apacheds-all jar.

This commit updates the spring-ldap-test project to remove its
apacheds-all dependency, replacing it with the individual ApacheDS
modules that provide the classes and functionality upon which it
actually depends.