LDAP-334: Support Spring Platform 2.0
Use the latest version of the Spring IO Plugin

Spring IO Platform 2.0 will remove the managed versions .properties
file as support for it has been removed in Spring Boot 1.3.

This commit moves the build onto a new version of the Spring IO Plugin
that uses the Maven bom rather than the properties file.