Upgrade to Spring IO Plugin 0.0.6.RELEASE
The latest version of the Spring IO Plugin removes the requirement for
code in src/test/java to be binary compatible with the dependency
versions defined by the Platform while still ensuring that the runtime
code is binary compatible.

This allows binary incompatible changes, such as those in AssertJ 2.6,
to be tolerated with no changes, as long as its only used in a
project's tests.