add missing delimiter line for source block in reference docs
The asciidoc source for the reference documentation was missing a block delimiter line
(four equals signs). Due to this, the table of of contents only listed headings up to
chapter 4.2. "Adding and Updating Data by Using DirContextAdapter". More importantly,
the formatting of the remaining document was messed up: Everything was wrapped in blocks
except the code examples.