Turn Off Automated Release
Spring LDAP is a small enough project that there are
a number of times when dependency upgrades and the like
are performed on the morning of the release.

At times, this get protracted for one reason or another,
and it is easy to forget that the release will automatically
kick off, sometimes while certain aspects of the release
are still being readied.

This commit turns of the scheduled release.